+++ adapters/folder/readFolderWithFIles_test.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/kr/pretty"
 	"github.com/simulot/immich-go/app"
-	"github.com/simulot/immich-go/helpers/configuration"
 	cliflags "github.com/simulot/immich-go/internal/cliFlags"
+	"github.com/simulot/immich-go/internal/configuration"
 	"github.com/simulot/immich-go/internal/exif"
 	"github.com/simulot/immich-go/internal/fileevent"
 	"github.com/simulot/immich-go/internal/filetypes"
