+++ app/cmd/_todo/duplicate/duplicate.nogo
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/simulot/immich-go/cmd"
-	"github.com/simulot/immich-go/helpers/gen"
+	"github.com/simulot/immich-go/internal/gen"
 	"github.com/simulot/immich-go/helpers/myflag"
 	"github.com/simulot/immich-go/immich"
 	"github.com/simulot/immich-go/ui"
