+++ app/cmd/immichSeverFlags.gono
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/simulot/immich-go/helpers/configuration"
+	"github.com/simulot/immich-go/internal/configuration"
 	"github.com/simulot/immich-go/immich"
 	"github.com/simulot/immich-go/internal/tzone"
 	"github.com/spf13/viper"
