+++ app/cmd/upload/upload_test.nogo
@@ -13,7 +13,7 @@ import (
 	"github.com/simulot/immich-go/browser"
 	"github.com/simulot/immich-go/cmd"
 	"github.com/simulot/immich-go/helpers/fileevent"
-	"github.com/simulot/immich-go/helpers/gen"
+	"github.com/simulot/immich-go/internal/gen"
 	"github.com/simulot/immich-go/immich"
 )
