+++ internal/fakefs/fakefs.go
@@ -11,8 +11,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/simulot/immich-go/helpers/gen"
 	"github.com/simulot/immich-go/internal/filenames"
+	"github.com/simulot/immich-go/internal/gen"
 )
 
 /*
