+++ internal/gen/slices.go
